westeros: perform more robust wayland event dispatching
In order to avoid potential lockups in handling Wayland protocol event
dispatching, adopt the same procedures that are used in
src/wayland/display.cpp. EventSource callbacks are enhanced:
- prepare() now dispatches pending events as long as there are
prepared events being provided,
- check() returns TRUE if new events can in fact be read,
- dispatch() only dispatches the pending events.

This is a speculative fix that attempts to address lockups that are
reported in issue #430 of the WPEWebKit repository.
https://github.com/WebPlatformForEmbedded/WPEWebKit/issues/430