wpeframework: compositorclient: adds a new wpe_compositorclient_get_parent_surface() api
This new api has a public header and is intended to be called from wpewebkit gstreamer
backend to get the native surface of the browser (in RDK systems this is the surface
that has been created by Thunder/WPEframework compositorclient code.

The api is used by GStreamer backend in wpewebkit to get the parent surface and then
set an external surface to waylandsink when it is configured to be used as the videoSink.

Setting an external video surface to waylandsink is required to get the video
surface set as a sub-surface of the browser surface.

v2: Use Native() api to obtain EGLNativeWindow and extract wl_surface from it rather
than adding new API to compositorClient to return the wl_surface.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>